+++ tests/frontend_laufzeit_test.js
@@ -12,6 +12,7 @@ function fakeEl(tag) {
     querySelectorAll(){ return []; }, remove(){}, open: false };
 }
 global.document = {
+  documentElement: { style: { setProperty(){} } },
   querySelector: (sel) => elemente[sel] || (elemente[sel] = fakeEl('div')),
   querySelectorAll: () => [],
   createElement: (t) => fakeEl(t),